improved detectionknn code

+++ birdwatcher/coordinatearrays.py
@@ -48,14 +48,14 @@ class CoordinateArrays(RaggedArray):
         return Path(apath)
 
 
-def create_coordarray(path, videofile,  metadata=None, overwrite=True):
+def create_coordarray(path, videofile, metadata=None, overwrite=True):
 
     if metadata is None:
         metadata = {}
     metadata.update(videofile.get_properties(affix='video_'))
     metadata.update({'birdwatcher_version': get_versions()['version']})
     coords = create_raggedarray(path, atom=(2,), metadata=metadata,
-                                overwrite=overwrite)
+                                overwrite=overwrite, dtype='uint16')
     return CoordinateArrays(coords, accessmode='r+')
